refactor(dashboard): extract the append-only telemetry series to a mixin (#1369)
Moves the six append-only telemetry accessors -- add/get_xvb_history,
add/get_network_history, add/get_disk_growth -- out of storage_service.py
into service/telemetry_store.py as TelemetryStoreMixin, which StateManager
inherits. Behaviour-identical: the 134 moved lines are byte-identical and the
full suite collects the same 2129 test IDs before and after.

atomic operations need the single DB-handle class and a transaction-manager
facade is design work. A mixin falls outside that reason -- same class, same
instance, same self, same connection, same transaction scope, and no new seam
between callers and the handle. The narrowing was ruled on #1367.

The boundary is the one the code already drew at _TELEMETRY_TABLES. Three of
those five tables move; blocks and worker_history stay, because worker_history
is read by get_worker_hashrate_by_config, which is worker-config domain and
the subject of the #1369 fix this unblocks.

Frees 121 lines of headroom under storage_service.py's budget ceiling
(1603 -> 1482), which the #1369 fix needs; that fix ships separately.
