+++ docs/dev/file-budget.tsv
@@ -8,7 +8,7 @@ dashboard/mining_dashboard/service/control_service.py	414
 dashboard/mining_dashboard/service/data_helpers.py	431
 dashboard/mining_dashboard/service/data_service.py	1454
 dashboard/mining_dashboard/service/egress.py	424
-dashboard/mining_dashboard/service/storage_service.py	1603
+dashboard/mining_dashboard/service/storage_service.py	1482
 dashboard/mining_dashboard/service/telegram_commands.py	1023
 dashboard/mining_dashboard/sim/donation_model.py	539
 dashboard/mining_dashboard/web/infra_views.py	507
